consistent types

+++ libs/observers2/nsupers_observer.hpp
@@ -48,7 +48,7 @@ struct NsupersFunc {
 /* returns WriteGridboxToArray which writes the number of superdrops in each
 gridbox to an array in a dataset in a store */
 template <typename Store>
-WriteGridboxToArray<Store> auto NsupersWriter(Dataset<Store> &dataset, const int maxchunk,
+WriteGridboxToArray<Store> auto NsupersWriter(Dataset<Store> &dataset, const size_t maxchunk,
                                               const size_t ngbxs) {
   return GenericWriteGridboxToXarray<Store, uint32_t, NsupersFunc>(dataset, "nsupers", "", "<u4", 1,
                                                                    maxchunk, ngbxs, NsupersFunc{});
